# Implement fANOVA using `sklearn` instead of `fanova`.