--- a/optuna/importance/_fanova/__init__.py
+++ b/optuna/importance/_fanova/__init__.py
@@ -1,0 +1,1 @@
+from optuna.importance._fanova._evaluator import FanovaImportanceEvaluator  # NOQA