--- a/setup.py
+++ b/setup.py
@@ -98,7 +98,6 @@ def get_extras_require() -> Dict[str, List[str]]:
             "chainer>=5.0.0",
             "cma",
             "fakeredis",
-            "fanova",
             "lightgbm",
             "mlflow",
             "mpi4py",