@HideakiImamura 2020-04-14
May I ask about the wording of the documentation rather than the contents of the code? 

Why should we use the `RandomSampler` with `FanovaImportanceEvaluator` as recommended in the document: "`It is recommended that trials are not skewed, that is, preferably belonging to a study that was optimized using the `RandomSampler`"? 
Please correct me if I'm wrong but the importance of hyperparameters can be judged by only using observed trial values according to Hutter et al. 2014 (Section 3.3). I think that how hyperparameters are selected is not matter.