@hvy 2020-04-14
You might be right as if samples are concentrated near a point, the leafs around this point should be proportionally weighted to their domains https://github.com/optuna/optuna/pull/1106/files#diff-34d5b18389a9c4d35ba97492b7dbc5e6R130-R133. At least, we don't have to mention this in the documentation as the logic mimics the paper. However, I would still expect some degradation as the forest will be less accurate when predicting the leaf midpoints where data points are scarce.