@hvy 2020-04-16
Noticed and fixed a bug with categorical midpoints traversals. Now the `_FanovaTree` doesn't care whether features are categorical or not. Features are one-hot encoded in the forest and the `_FanovaTree` will process the columns as if they were any other numerical column. 

<details>
  <summary>LightGBM, 100 trials, `RandomSampler`</summary>
  <img src="https://user-images.githubusercontent.com/5983694/79442468-89a13980-8013-11ea-9a89-4ea97a315e7c.png">
</details>
<details>
  <summary>PyTorch, 100 trials, `RandomSampler`</summary>
  <img src="https://user-images.githubusercontent.com/5983694/79442587-b6555100-8013-11ea-8ae1-828ac93e2db6.png">
</details>
<details>
  <summary>ffmpeg, 150 trials, `RandomSampler`</summary>
  <img src="https://user-images.githubusercontent.com/5983694/79442735-e997e000-8013-11ea-8549-e890522eda32.png">
</details>