@HideakiImamura 2020-04-17
Thank you for updating. I checked the changed codes and understood the behavior of`_FanovaTree`, which does not care whether features are categorical or not. LGMTM.