@hvy 2020-04-25
The last commit https://github.com/optuna/optuna/pull/1106/commits/abed879915a5c2d8be24f1c107593b7362f0edbf has `_WeightedRunningStats` removed and uses `numpy.ndarray`s instead . 

Evaluating single and pairwise importances with my PyTorch example (including several categorical features which are more computationally demanding in Optuna's implementation due to the one-hot sparse encoding) is now roughly twice as fast. 

#### Before 

```
_Fanova: 1.581s 
fANOVA: 0.1682s 
```
_Average time taken per feature(s) (single and pairwise). 
`_Fanova` is Optuna's, `fANOVA` is AutoML's._

#### After 
```
_Fanova: 0.8128s
fANOVA: 0.1648s 
```

_Verified that evaluated importances are the same as before (https://github.com/optuna/optuna/pull/1106#issuecomment-614544840)._

#### Note

I think further performance improvements require e.g. changes to the marginal prediction tree traversal as that's now the most significant bottleneck associated with the Python runtime overhead. Also cleaned up some interfaces and implementation details.