@hvy 2020-07-02
Resolved conflicts and added references to the original work. PTAL.